Added new messages
LaserSimpleScan message only contains the minimum
distance measured by the laser and the respective ray.
SwarmAgentData contains info about each swarm agent
in the team. The rel_position is relative to the
leader.
SwarmLeaderData.msg contains info about the leader of
the team. The agents field contains the info about
each agent (the agents[idx].rel_position is the
relative position to the agent).